% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute.R
\name{ChartAttribute}
\alias{ChartAttribute}
\title{Chart attribute specification}
\usage{
ChartAttribute(
  name,
  data = NULL,
  displayName = name,
  text = NULL,
  missing_value = NA,
  hrefBase = NULL,
  href = NULL,
  max_digits = NULL,
  dataset_summary_fn = NULL,
  summary_fn = NULL,
  allow_extra_data = FALSE,
  missing_fill = NA,
  only_ranks = NULL
)
}
\arguments{
\item{name}{A name or label. In case of generating charts from a tree structure
(nested lists), \code{name} identifies the attribute.}

\item{data}{The attribute data, which can be a vector or matrix/data frame
(see details)}

\item{displayName}{Optional longer label displayed next to the attribute value
instead of the \code{name}}

\item{text}{Text displayed for a node, or a function used to generate the text,
in the form \code{text(node_name, attribute_value)}}

\item{missing_value}{Behavior in case of missing attribute values:
By default (\code{missing_value = NA}), the attribute is hidden from the list
of attributes in the top-right corner if the underlying data is \code{NA} for
the currently selected node.
To explicitly always include the attribute in the list, e.g. set
\code{missing_value = 'N/A'}.
\emph{Note}: If using this attribute for \emph{coloring}, \code{missing_value} should also
be set to some character string, otherwise the color for the value 0 is assumed.
\code{make_krona(color = ...)} does this automatically.}

\item{hrefBase}{Optional base URL is prepended to each value (or \code{href} path).
The text is turned into a link if \code{hrefBase} is defined.}

\item{href}{URL or URL component following the common \code{hrefBase} prefix,
or a function for generating it in the form \code{href(node_name, attribute_value)}}

\item{max_digits}{The number of digits to use when formatting numeric data.
The default is \code{max_digits} from \link{ChartOpts}.}

\item{summary_fn}{Summary function for aggregating \code{data} for higher ranks.
The function should look as follows: \code{summary_fn(attribute_data, magnitudes)}
where \code{attribute_data} is a vector of values.
The default function for numeric data calculates a weighted mean:
\code{weighted.mean(attribute_data, magnitudes, na.rm=TRUE)}.
In case of a \emph{character} vector/matrix, the default function assigns an item
to a lower/inner rank if all nested items are equal, otherwise it assigns \code{NA}.}

\item{allow_extra_data}{Suppress warnings regarding extra entries in \code{data}
\emph{not} found in the classification.}

\item{missing_fill}{Numeric or character value to use for missing entries that
are present in the \code{classification} but not in the attribute \code{data}.
Default: \code{NA} (another good choice for numeric values could be missing_fill = 0).
Not to be mistaken with \code{missing_value}, which just controls how attributes
are displayed.}

\item{only_ranks}{Optional vector of names or indices of \code{classification}
rank columns for which the attribute should be displayed}

\item{dataset_summary_fn:}{Function for a rowwise summary of \code{data} columns across
datasets (see \code{groups} in \link{make_krona}). The function must accept a vector
vector of values and return a single value.
The default is \code{sum} for numeric values. Character items from a group are
only retained if are all equal, otherwise \code{NA} is returned.}
}
\description{
Attributes provide information about nodes of in the classification tree,
which is displayed in the top-right area of the chart once a node is clicked.
If generating Krona charts from a matrix/data frame/phyloseq object, both
\code{name} and \code{data} are required.

If a pre-computed nested tree is provided as \code{classification}, only \code{name}
is required to indicate the name of the list entry containing the data.
Also, some other options don't apply: \code{dataset_summary_fn}, \code{allow_extra_data},
\code{missing_fill}, \code{only_ranks}.

Numeric data can be used for custom coloring (see \code{color} in \link{make_krona}).
External links can be displayed by setting \code{hrefBase} and \code{href}.
Text data may contain HTML code (automatically escaped).
}
\details{
Attributes reflect the structure of the Krona XML format. Even the abundance
information (magnitude) is stored as an XML attribute (\link{make_krona.matrix} uses
the name \code{n} for that).
\subsection{How does \code{data} look like?}{

It may be provided as:
\itemize{
\item A vector with entries corresponding to the different classes in the hierarchy;
values cannot vary across different datasets, if there are any
\item A matrix or data frame with (optionally named) columns corresponding to data sets
(see \code{dataset_group} in \link{make_krona}). If the number of columns corresponds to the
columns of the \code{magnitude} supplied to \link{make_krona}, values are aggregated
by \code{dataset_group} with \code{dataset_summary_fn}.
}

If if vector names or matrix row names are set both in the \code{make_krona}
classification \emph{and} the attribute \code{data}, they will be used for matching the two.
There may be missing values in \code{data} (see also \code{missing_fill}).
If there are no row names, \code{classification} and attribute \code{data} entries
must align.

Numeric data is aggregated with \code{summary_fn} (default: average).
Text data is usually only available in the lowest rank (outermost level
of the hierarchy) unless it is identical across all children of a node,
in which case it gets propagated upwards. The behavior can be changed with
\code{summary_fn}.
}

\subsection{Limitations}{

Attribute lists are not supported
(as in http://marbl.github.io/Krona/examples/metarep-ec.krona.html)
}
}
