% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{ChartDisplayOpts}
\alias{ChartDisplayOpts}
\title{Configure the initial chart appearance}
\usage{
ChartDisplayOpts(
  collapse = getOption("taxplore.display.collapse", TRUE),
  showMagnitude = getOption("taxplore.display.showMagnitude", FALSE),
  color = getOption("taxplore.display.color"),
  depth = getOption("taxplore.display.depth"),
  font = getOption("taxplore.display.font", 11),
  key = getOption("taxplore.display.key", TRUE),
  dataset = getOption("taxplore.display.dataset"),
  node = getOption("taxplore.display.node")
)
}
\arguments{
\item{collapse}{(logical or 'true'/'false') Should the chart be simplified,
collapsing 'redundant' wedges? (on by default)}

\item{showMagnitude}{(logical or 'true'/'false') Should magnitudes be displayed
in [brackets]?}

\item{color}{(logical or 'true'/'false') Color by a custom scheme along a
a gradient of an attribute variable (see \code{color}). Automatically activated
\code{TRUE} if \code{color} is set in \link{make_krona} or \link{plot_krona}, unless explicitly
set to \code{FALSE} here.}

\item{depth}{(number) Maximum number of ranks/levels that should be displayed
(unlimited by default)}

\item{font}{(number) Font size (px/pt) (default: 11)}

\item{key}{(logical or 'true'/'false') Should a list of small wedges
('key') be displayed on the bottom right of the chart?
(displayed by default)}

\item{dataset}{(number) Dataset number (see \code{group}/\code{group_col})}

\item{node}{(number) Node number to highlight...}
}
\description{
Configure the initial chart appearance
}
\details{
The \code{ChartDisplayOpts} configure the initial (default) appearance of embedded Krona
charts. Some of the settings correspond to checkboxes on the left.
The same parameters can also be passed as key-value (GET) URL parameters
to Krona HTML charts, see Krona wiki:
\itemize{
\item https://github.com/marbl/Krona/wiki/Customizing-the-view
\item https://github.com/marbl/Krona/wiki/Changing-chart-viewing-defaults
}

There are different ways to apply these options:
\itemize{
\item Directly supply them to \link{make_krona} as follows:
\code{make_krona(..., display = ChartDisplayOpts(...))}
\item Set global defaults before the call(s) to \code{make_krona()}:
\code{set_chart_display_opts(...)}
\item Set with \link{options} as follows: \code{options(taxplore.display.optionname = 'value')}
}
}
\seealso{
\link{make_krona}
}
