% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{ChartOpts}
\alias{ChartOpts}
\title{General chart options}
\usage{
ChartOpts(
  max_digits = getOption("taxplore.max_digits", 7),
  hue_range = getOption("taxplore.hue_range", c(0, 120)),
  unknown_label = getOption("taxplore.unknown_label"),
  root_label = getOption("taxplore.root_label", "Root"),
  total_label = getOption("taxplore.total_label", "Total"),
  method = getOption("taxplore.method"),
  resources_url = getOption("taxplore.resources_url"),
  interactive = getOption("taxplore.interactive"),
  minify = getOption("taxplore.minify"),
  kronatools_dir = getOption("taxplore.kronatools_dir")
)
}
\arguments{
\item{max_digits}{Maximum number of significant digits to print for decimal numbers.
Reducing to <7 digits (the default) may help in obtaining smaller HTML files.}

\item{total_label}{Label for the total abundance displayed on the top right
of the chart.}

\item{method}{HTML generation method ('internal' or 'kronatools'),
'  passed to \link{generate_krona_html}}

\item{minify}{If \code{TRUE}, a minified version of the Krona JavaScript is included
Automatically activated in \link{plot_krona} when embedded
This option is ignored if \code{method = 'kronatools'} and has no effect
if \code{resources_url} is provided.}

\item{hue_range:}{Hue at the start and end points of the fill gradient if coloring
by a user-specified gradient (see \code{color} and \code{color_value_range} arguments
in \code{make_krona}).}

\item{unknown_label:}{Label to set instead of \code{NA}s in the classification/taxonomy
(if any present).
By default, \code{NA}s are left in place, resulting in fewer child nodes. Krona
charts handle this situation well (see e.g. https://krona.sourceforge.net/examples/xml.krona.html).}

\item{root_label:}{Label to use for the root of the taxonomic tree. It is
only displayed if the uppermost level in the hierarchy has more than one
category, or \code{collapse} is disabled (see \link{ChartDisplayOpts}).}

\item{resources_url:}{Optional URL pointing to remote Javascript and image resources,
which will then be linked to instead of embedded into the chart.
The resulting charts are not be standalone and require an internet connection
for displaying, but the HTML file size is smaller.
Example: \code{resources_url = 'http://marbl.github.io/Krona/'}
The given URL address should serve the contents of the 'src'
and 'img' directories from KronaTools (https://github.com/marbl/Krona/tree/master/KronaTools)
(see \link{copy_krona_resources}).}

\item{kronatools_dir:}{Optional path pointing to a 'KronaTools' directory
extracted from the Krona source code archive, downloaded from
\href{https://github.com/marbl/Krona/releases/latest}{here}.
This option is ignored if \code{method = 'internal'}.}

\item{snapshot}{If \code{TRUE}, the Krona chart is immediately turned into a non-interactive
snapshot. In contrast to the \code{interactive} parameter of \code{plot_krona},
this option permanently modifies the chart HTML code, so files exported with
\code{outfile = ...} will also be non-interactive.}
}
\description{
General chart options
}
\details{
There are different ways to apply these options:
\itemize{
\item Supply options directly to \link{make_krona} or \link{plot_krona}:
\code{plot_krona(..., root_label = 'Records')}
\item Supply as \code{opts} to \link{make_krona} or \link{plot_krona}:
\code{make_krona(..., opts = ChartOpts(...))}
\item Set global defaults before the call(s) to \code{make_krona()}:
\code{set_chart_opts(...)}
\item Set with \link{options} as follows: \code{options(taxplore.optionname, value)}
}
}
\seealso{
\link{make_krona}
}
