% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_html.R
\name{copy_krona_resources}
\alias{copy_krona_resources}
\title{Copy chart resources to a specified directory}
\usage{
copy_krona_resources(dest_dir)
}
\arguments{
\item{dest_dir}{Target directory, which can be linked by multiple charts
(see \code{resources_url} in \code{\link[=ChartOpts]{ChartOpts()}})}
}
\description{
The images and JavasScript files are unmodified snapshots of the original files
from the \href{https://github.com/marbl/Krona}{Krona} repository
(master branch, commit d1479b3).
}
