% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_html.R
\name{generate_krona_html}
\alias{generate_krona_html}
\title{Generates a HTML chart from a character string containing the Krona XML.}
\usage{
generate_krona_html(xml, method = c("internal", "kronatools"), ...)
}
\description{
With \code{method = 'internal'} (the default), R implementation in this package is used,
which by default embeds the unmodified Krona JavaScript source code delivered with
this package, or a minified version of it with \code{minify = TRUE}.
(see \code{\link[=generate_krona_html_internal]{generate_krona_html_internal()}}).
}
\details{
Alternatively, \code{method = 'kronatools'} uses the official Krona software
(see \code{\link[=kt_import_xml]{kt_import_xml()}}).
The resulting HTML should be the same, although the 'internal' method provides
additional options that are useful when embedding charts in documents.
Therefore, this is the default method.
}
