% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_html.R
\name{generate_krona_html_internal}
\alias{generate_krona_html_internal}
\title{Generate a HTML chart in R}
\usage{
generate_krona_html_internal(
  xml,
  outfile = NULL,
  kronatools_dir = NULL,
  resources_url = NULL,
  minify = FALSE,
  snapshot = FALSE,
  ...
)
}
\arguments{
\item{xml}{Krona XML string}

\item{outfile}{Write the string to file instead of returning the HTML character
string.}

\item{resources_url, interactive, minify, snapshot}{See \link{ChartOpts}}

\item{...}{Additional arguments are currently ignored}
}
\value{
The generated HTML, or NULL if \code{outfile} was specified
}
\description{
This does not rely on the external KronaTools software (see \link{kt_import_xml}).
}
