% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grasslandfungi}
\alias{grasslandfungi}
\alias{grasslandfungi.records}
\title{Grassland fungi survey data}
\format{
An object of class \code{data.frame} with 54 rows and 13 columns.

An object of class \code{data.frame} with 127 rows and 7 columns.
}
\source{
\url{https://swissfungi.wsl.ch}
}
\usage{
grasslandfungi

grasslandfungi.records
}
\description{
Dataset from a grassland plot in the Eastern Swiss mountains (1'500 m elevation)
where a total of 54 occurrences of macrofungi were recorded during two visits
in September 2021.
}
\details{
The data frames contain the taxonomic lineages (one row per taxon) with
associated record counts during the two visits (\code{visit_1} and \code{visit_2}).

\code{grasslandfungi.records} simply contains the full taxonomic lineages for
\emph{every} record (taxa repeated), without any additional information.

\code{grasslandfungi} has \emph{one row per taxon}, along with the record counts
from the two surveys (\code{visit_1} and \code{visit_2}) and the total record count
(\code{n_total}). Further, there are \emph{GBIF} and \emph{SwissFungi} taxon IDs and
the IUCN redlist status according to the current Swiss red list of macrofungi.
}
\keyword{datasets}
