% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_html.R
\name{kt_import_xml}
\alias{kt_import_xml}
\title{Generate a HTML chart using the official Krona tools}
\usage{
kt_import_xml(
  xml,
  outfile = NULL,
  kronatools_dir = NULL,
  resources_url = NULL,
  ...
)
}
\arguments{
\item{xml}{Krona XML string}

\item{outfile}{Write the string to file instead of returning the HTML character
string.}

\item{kronatools_dir, resources_url}{see \link{ChartOpts}}

\item{...}{Additional arguments are ignored}
}
\value{
The generated HTML, or NULL if \code{outfile} was specified
}
\description{
The \code{ktImportXML} script from the Krona tools  (\url{https://github.com/marbl/Krona})
must be available in \emph{$PATH}.
Alternatively, the path to the extracted \emph{KronaTools} directory can be
specified with \code{kronatools_dir}.
}
