% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_krona.R
\name{make_krona}
\alias{make_krona}
\alias{make_krona.taxonomyTable}
\alias{make_krona.phyloseq}
\alias{make_krona.data.frame}
\alias{make_krona.matrix}
\alias{make_krona.list}
\title{Generate a chart}
\usage{
make_krona(x, ...)

\method{make_krona}{taxonomyTable}(x, tax_ranks = NULL, ...)

\method{make_krona}{phyloseq}(
  x,
  magnitude = NULL,
  outfile = NULL,
  tax_ranks = NULL,
  dataset_group = NULL,
  group_vars = NULL,
  color = NULL,
  color_col = NULL,
  color_label = NULL,
  group_sep = " ",
  shorten_group = function(x) abbreviate(x, 60),
  ...
)

\method{make_krona}{data.frame}(classification, ...)

\method{make_krona}{matrix}(
  x,
  magnitude = NULL,
  outfile = NULL,
  dataset_group = NULL,
  color = NULL,
  color_label = NULL,
  color_value_range = NULL,
  attributes = NULL,
  dataset_summary_fn = sum,
  display = NULL,
  method = NULL,
  opts = NULL,
  ...
)

\method{make_krona}{list}(
  x,
  magnitude = "n",
  outfile = NULL,
  datasets = NULL,
  attributes = NULL,
  color = NULL,
  color_value_range = NULL,
  opts = NULL,
  display = NULL,
  ...
)
}
\arguments{
\item{x}{A classification matrix/data frame containing a hierarchical classification
(e.g. taxonomic lineages), a phyloseq object or a tree data structure
made of nested lists (see \emph{details} below)}

\item{...}{Passed on to \code{make_krona.matrix}, and remaining arguments are ultimately
passed as \link{ChartOpts} as an alternative to \code{opts} for configuring HTML generation.}

\item{tax_ranks}{Character vector with phyloseq taxonomy ranks names that
should be chosen from the taxonomy table (default: all ranks)}

\item{magnitude}{Specifies the magnitudes (abundances) of the different groups
in the classification. If missing, equal magnitudes are assumed for all entries.
Either a vector or matrix/data frame (see \emph{details})}

\item{outfile}{Optional output file. If not provided, the HTML output is returned
as character string.}

\item{dataset_group}{Optional character vector specifying the grouping of
magnitude columns or phyloseq samples into distinct datasets.
By default, data in different columns of matrix-like abundance objects are
summed up by row and globally displayed as one dataset.
The \code{dataset_group} vector must be of length \code{ncol(magnitude)}.
The special value \emph{"separate"} assumes that all all columns/phyloseq samples
are separate data sets.
Adjust the by-row summary function with \code{dataset_summary_fn}.}

\item{group_vars, group_sep, shorten_group}{Character vector with phyloseq sample
variables that define the grouping into different datasets (see \code{dataset_group}.
Multiple variables are joined together using \code{group_sep} to form the \code{dataset_group}
vector.
Overly long strings are further shortened using the \code{shorten_group} function.}

\item{color_col}{Name of a column in the \emph{phyloseq} taxonomy table, which
contains numeric values used for coloring (see \code{color}).
\emph{Note} that it is rather unusual to store extra (numeric) data in the
taxonomy table, but it is still possible.}

\item{color_label}{Label of the color scale (bottom left).
Must be supplied if \code{color} is a data vector/matrix.
Otherwise, the \code{name} or \code{displayName} of the \link{ChartAttribute} is used.}

\item{color_value_range}{Lower and upper value limits that map to the color
range (see \code{hue_range} in \link{ChartOpts}). By default, the full \code{range()}
is used.}

\item{attributes}{Optional list of \link{ChartAttribute} providing more information
about nodes of in chart (when clicked), or for the coloring.}

\item{dataset_summary_fn}{Function used to summarize the magnitudes from several
samples (see \code{dataset_group}). \code{sum} is the default, but \code{mean} is also
a good choice.}

\item{display}{Optional \link{ChartDisplayOpts} configuring the initial
appearance of the charts when embedded in RStudio or rendered documents.
See \link{set_chart_display_opts} to apply these settings globally for multiple
charts.}

\item{opts}{Optional \link{ChartOpts} affecting how HTML charts are generated
See also \link{set_chart_opts} to apply these settings globally for multiple charts.}

\item{datasets}{If \code{x} is a tree data structure with magnitudes for >1 dataset
(stored in 'n' attribute by default), then the names of these data sets
need to be provided with \code{datasets} (character vector).}

\item{color:}{Indicate data to be used for custom coloring of the chart.
Should be one of:
\itemize{
\item Name of a \link{ChartAttribute} provided in the \code{attributes} list, which
has \emph{numeric} data to color by.
\item \emph{Numeric vector/matrix} that directly provides the necessary data.
See \link{ChartAttribute} for information about the data format and how values
are aggregated within higher classification ranks.
\code{color_label} also needs to be specified.
}}
}
\value{
Krona chart HTML code as character vector
}
\description{
This method generates the HTML code for a chart, which it returns or writes to a file.
}
\section{Matrix-like classification}{
The classification (\code{x}) must be one of:
\itemize{
\item A \emph{matrix} or \emph{data frame} where each row defines a lineage of the
hierarchy from left (higher ranks = inner circles in chart) to right
(lower ranks = outer circles).
Usually, an associated \code{magnitudes} vector or matrix/data frame is supplied
along, indicating the size (weight) of the nodes in the chart.
If not provided, equal weights are given to all nodes.
\item A \emph{Phyloseq object} with at least a \code{tax_table} and \code{otu_table} component,
or a simple \code{taxonomyTable} with or without additional \code{magnitudes}.
Grouping into datasets can be done with \code{group_vars} (requires \code{sample_data}
to be present).
}

\code{magnitudes} modify the width of the outermost nodes in the hierarchy
(and the inner nodes along with them).
It should be a vector of \code{nrow(x)} or a matrix/data frame matching the rows of \code{x}.
By default, magnitude values from the different columns are summed up by row
(configurable with \code{dataset_summary_fn}).
Optionally samples can be grouped into datasets (see \code{dataset_group}).
Any vector names/row names (if present) must match with nammes of \code{x}.
\subsection{Phyloseq objects}{

If \code{x} is a \emph{Phyloseq object}, no separate \code{magnitude} is required as abundances
are taken from the \code{otu_table()} slot (but \code{magnitude = ...} can override it).
Set \code{magnitude = FALSE} to give equal weight to all taxa.
}
}

\section{Tree data structure}{
\code{x} can be a pre-assembled \emph{tree} made of nested lists, which is written to
the Krona XML unmodified.
Each node requires at least a \code{name} attribute.
The optional magnitude attribute (default name: \code{magnitude = 'n'}) specifies
the node weights for one or multiple datasets (specified with \code{datasets}).
Additional data can be provided (see \code{attributes}).
}

\seealso{
\link{plot_krona}, \link{ChartOpts}, \link{ChartDisplayOpts}
}
