% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{plot_krona}
\alias{plot_krona}
\title{Display a Krona chart as HTML widget}
\usage{
plot_krona(
  x,
  ...,
  file = NULL,
  outfile = NULL,
  opts = NULL,
  display = NULL,
  interactive = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{x}{Character string with a pre-generated HTML chart, or a
matrix, data frame, list or phyloseq object that serves as input for
generating the chart with \code{\link[=make_krona]{make_krona()}}.}

\item{...}{Additional arguments passed to \link{make_krona}}

\item{file}{File path, URL or file connection pointing to an already generated
chart (passed instead of \code{x})}

\item{outfile}{Write the embedded chart to file}

\item{display}{A \code{\link[=ChartDisplayOpts]{ChartDisplayOpts()}} instance configuring the (initial)
appearance of embedded charts or snapshots.}

\item{interactive}{If \code{FALSE}, a non-interactive HTML chart is shown,
which does not have any buttons and does not allow selecting and zooming.
This is primarily useful for taking snapshots of embedded widgets,
and is automatically activated in \link{plot_krona} when rendering in static documents.
If set to \code{TRUE}, charts embedded in static documents such as PDF and Word
are unmodified snapshots with buttons and other controls.
Note that \code{outfile = ...} still exports a "normal" interactive Krona chart
(see \code{snapshot} option in \link{ChartOpts} for making the exported chart non-interactive).}

\item{width, height}{Width/height of the widget.
Must be a valid CSS unit (like \code{'100\%'}, \code{'400px'}, \code{'auto'}) or a number,
which will be coerced to a string and have \code{'px'} appended.}
}
\value{
A \href{https://www.htmlwidgets.org}{HTML widget} embedable in
\href{https://rmarkdown.rstudio.com}{R-Markdown} / \href{https://quarto.org}{Quarto}
documents.
Optionally,  \code{outfile} can be specified to also save the chart to a HTML file.
}
\description{
Pre-generated HTML charts can be displayed given a character string or
supplied with \code{file = ...}, or the chart can be directly generated with
\code{\link[=make_krona]{make_krona()}} and directly displayed.
}
\details{
When generating the widget, some settings are adjusted depending on the environment.
In static (non-HTML) documents such as PDF or Word, a snapshot is taken,
which hides the buttons, search box and other things (can be changed with
\code{interactive = TRUE}). Still, some additional configuration might be required
to ensure that the charts look nice in every context (see \code{\link[=taxplore_configure_snapshot]{taxplore_configure_snapshot()}}).
}
\seealso{
\code{\link[=make_krona]{make_krona()}}, \code{\link[=ChartOpts]{ChartOpts()}}, \code{\link[=ChartDisplayOpts]{ChartDisplayOpts()}}
}
