% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{set_chart_display_opts}
\alias{set_chart_display_opts}
\title{Set global chart display options to override the default \link{ChartDisplayOpts}.}
\usage{
set_chart_display_opts(...)
}
\arguments{
\item{...}{Named arguments}
}
\description{
The options are added to \link{options} using \code{options(taxplore.display.optionname = value)}.
The settings are used to set the initial appearance of all all subsequent
charts shown with \link{plot_krona}.
}
\examples{
set_chart_opts(hue_range = c(40, 140))

}
