% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{set_chart_opts}
\alias{set_chart_opts}
\title{Set global chart options to override the default \link{ChartOpts}.}
\usage{
set_chart_opts(...)
}
\arguments{
\item{...}{Named arguments}
}
\description{
The options are added to \link{options} using \code{options(taxplore.key = value)}.
The settings are used for all subsequent charts generated with \link{plot_krona} and \link{make_krona}.
}
\examples{
set_chart_opts(hue_range = c(40, 140))

}
