% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{taxplore_configure_snapshot}
\alias{taxplore_configure_snapshot}
\title{Set sensible Knitr defaults for chart snapshots}
\usage{
taxplore_configure_snapshot(dpi = 300, dev = "png", screenshot.delay = 0.05)
}
\arguments{
\item{screenshot.delay}{Time (in s) to wait before taking a screenshot of a
HTML widget (such as the Krona charts from this package, but not exclusively).
The default of 0.05s differs from the default 0.2s used by Knitr. To be precise,
the Krona charts need no delay at all.
In case of problems with other widgets, a longer delay may be set.}

\item{document.dpi}{Resolution of images in static non-HTML (e.g. PDF) documents}

\item{document.dev}{Snapshot format for non-HTML documents.
The default is to generate PNG images, as embedding them as vector graphics
may currently not result in the correct dimensions due to a
\href{https://github.com/quarto-dev/quarto-cli/issues/7682}{problem with in Webshot2}.}
}
\description{
Sets default chunk options that work well both with HTML documents and other
formats such as PDF.
\strong{Caution}: The default image format and resolution are set for \emph{all chunks},
not just those containing Krona charts.
}
